+++ packages/propeller/src/StreamsSink.test.mts
@@ -2,6 +2,7 @@ import { describe, test, expect, vi } from "vitest"
 import { BatchLimiter, StreamsSink } from "./StreamsSink.mjs"
 import { ITimelineEvent, StreamId, StreamName } from "@equinox-js/core"
 import { IngesterBatch } from "./Types.js"
+import { StreamResult } from "./Sinks.js"
 
 const mkEvent = (type: string, index: bigint): ITimelineEvent => ({
   id: "",
@@ -12,13 +13,13 @@ const mkEvent = (type: string, index: bigint): ITimelineEvent => ({
   index,
 })
 
-const mkBatch = (onComplete: () => void, index?: bigint): IngesterBatch => ({
+const mkBatch = (onComplete: () => void, index?: bigint, checkpoint = 5n): IngesterBatch => ({
   items: Array.from({ length: 10 }).map((_, i): [StreamName, ITimelineEvent] => [
     StreamName.create("Cat", StreamId.create("stream" + i)),
     mkEvent("Something", index ?? BigInt(i)),
   ]),
   isTail: false,
-  checkpoint: 5n,
+  checkpoint,
   onComplete,
 })
 
@@ -29,12 +30,14 @@ describe("Concurrency", () => {
       const active = new Set<string>()
       let maxActive = 0
       const ctrl = new AbortController()
+
       async function handler(stream: StreamName) {
         active.add(stream)
         maxActive = Math.max(maxActive, active.size)
         await new Promise(setImmediate)
         active.delete(stream)
       }
+
       const limiter = new BatchLimiter(1)
       const sink = new StreamsSink(handler, concurrency, limiter)
       const sinkP = sink.start(ctrl.signal)
@@ -56,10 +59,12 @@ describe("Concurrency", () => {
 test("Correctly merges batches", async () => {
   let invocations = 0
   const ctrl = new AbortController()
+
   async function handler() {
     invocations++
     await new Promise(setImmediate)
   }
+
   const limiter = new BatchLimiter(3)
   const sink = new StreamsSink(handler, 10, limiter)
   const sinkP = sink.start(ctrl.signal)
@@ -92,10 +97,12 @@ const mkSingleBatch = (
 test("Correctly limits in-flight batches", async () => {
   let invocations = 0
   const ctrl = new AbortController()
+
   async function handler() {
     invocations++
     await new Promise((res) => setTimeout(res, 10))
   }
+
   const limiter = new BatchLimiter(3)
   const sink = new StreamsSink(handler, 1, limiter)
   const sinkP = sink.start(ctrl.signal)
@@ -127,6 +134,7 @@ test("Ensures at-most one handler is per stream", async () => {
   let invocations = 0
   let events = 0
   const ctrl = new AbortController()
+
   async function handler(sn: StreamName, evs: unknown[]) {
     events += evs.length
     invocations++
@@ -135,6 +143,7 @@ test("Ensures at-most one handler is per stream", async () => {
     await new Promise((res) => setTimeout(res, 10))
     active--
   }
+
   const limiter = new BatchLimiter(10)
   const sink = new StreamsSink(handler, 100, limiter)
   const sinkP = sink.start(ctrl.signal)
@@ -159,4 +168,169 @@ test("Ensures at-most one handler is per stream", async () => {
   // onComplete is called in order and for every batch
   expect(completed.mock.calls).toEqual(Array.from({ length: count }).map((_, i) => [BigInt(i)]))
   await sinkP
+})
+
+test("Old events are ignored even if re-submitted", async () => {
+  let invocations = 0
+  const ctrl = new AbortController()
+
+  async function handler() {
+    invocations++
+    await new Promise(setImmediate)
+  }
+
+  const limiter = new BatchLimiter(3)
+  const sink = new StreamsSink(handler, 10, limiter)
+  const sinkP = sink.start(ctrl.signal)
+
+  const checkpoint = vi.fn().mockResolvedValue(undefined)
+
+  // each mkBatch has 1 event in 10 streams
+  await sink.pump(mkBatch(checkpoint, 0n, 0n), ctrl.signal)
+  await sink.pump(mkBatch(checkpoint, 1n, 1n), ctrl.signal)
+
+  await limiter.waitForEmpty()
+  expect(invocations).toBe(10)
+
+  // whoopsie, at-least-once delivery means we can receive older events again
+  await sink.pump(mkBatch(checkpoint, 0n, 2n), ctrl.signal)
+  await sink.pump(mkBatch(checkpoint, 1n, 3n), ctrl.signal)
+  await limiter.waitForEmpty()
+  ctrl.abort()
+
+  expect(invocations).toBe(10)
+  expect(checkpoint).toHaveBeenCalledTimes(4)
+  await sinkP
+})
+
+describe("StreamResult", () => {
+  test("StreamResult.PartiallyProcessed Retries the stream with the rest of the events", async () => {
+    let invocations = 0
+    const eventCount: number[] = []
+    const ctrl = new AbortController()
+
+    async function handler(sn: StreamName, events: ITimelineEvent[]) {
+      invocations++
+      eventCount.push(events.length)
+      await new Promise(setImmediate)
+      return StreamResult.PartiallyProcessed(2)
+    }
+
+    const limiter = new BatchLimiter(3)
+    const sink = new StreamsSink(handler, 10, limiter)
+    const sinkP = sink.start(ctrl.signal)
+
+    const checkpoint = vi.fn().mockResolvedValue(undefined)
+
+    const batch: IngesterBatch = {
+      items: Array.from({ length: 10 }, (_, i) => [
+        StreamName.parse("Cat-stream0"),
+        mkEvent("Something", BigInt(i)),
+      ]),
+      isTail: false,
+      checkpoint: 0n,
+      onComplete: checkpoint,
+    }
+
+    // each mkBatch has 1 event in 10 streams
+    await sink.pump(batch, ctrl.signal)
+
+    await limiter.waitForEmpty()
+    ctrl.abort()
+
+    expect(invocations).toBe(5)
+    expect(eventCount).toEqual([10, 8, 6, 4, 2])
+    expect(checkpoint).toHaveBeenCalledTimes(1)
+    await sinkP
+  })
+
+  test("OverrideNextIndex ignores intervening events", async () => {
+    let invocations = 0
+    const eventIndices: bigint[] = []
+    const ctrl = new AbortController()
+
+    async function handler(sn: StreamName, events: ITimelineEvent[]) {
+      invocations++
+      for (const event of events) eventIndices.push(event.index)
+      await new Promise(setImmediate)
+      const ver = events[0].index + BigInt(events.length)
+      if (ver < 20) return StreamResult.OverrideNextIndex(20n)
+      return
+    }
+
+    const limiter = new BatchLimiter(3)
+    const sink = new StreamsSink(handler, 10, limiter)
+    const sinkP = sink.start(ctrl.signal)
+
+    const checkpoint = vi.fn().mockResolvedValue(undefined)
+
+    const batch = (from: number, count = 10): IngesterBatch => ({
+      items: Array.from({ length: count }, (_, i) => [
+        StreamName.parse("Cat-stream0"),
+        mkEvent("Something", BigInt(from + i)),
+      ]),
+      isTail: false,
+      checkpoint: BigInt(from + count),
+      onComplete: checkpoint,
+    })
+
+    // each mkBatch has 1 event in 10 streams
+    await sink.pump(batch(0), ctrl.signal)
+    await limiter.waitForEmpty()
+    await sink.pump(batch(10), ctrl.signal)
+    await limiter.waitForEmpty()
+    await sink.pump(batch(20, 2), ctrl.signal)
+    await limiter.waitForEmpty()
+    ctrl.abort()
+
+    const expectedIndices = [0n, 1n, 2n, 3n, 4n, 5n, 6n, 7n, 8n, 9n, 20n, 21n]
+    expect(eventIndices).toEqual(expectedIndices)
+    expect(invocations).toBe(2)
+    expect(checkpoint).toHaveBeenCalledTimes(3)
+    await sinkP
+  })
+
+  test("NoneProcessed will retry the stream", async () => {
+    let invocations = 0
+    const eventIndices: bigint[] = []
+    const ctrl = new AbortController()
+
+    let failedOnce = false
+    async function handler(sn: StreamName, events: ITimelineEvent[]) {
+      invocations++
+      for (const event of events) eventIndices.push(event.index)
+      await new Promise(setImmediate)
+      if (!failedOnce) {
+        failedOnce = true
+        return StreamResult.NoneProcessed
+      }
+    }
+
+    const limiter = new BatchLimiter(3)
+    const sink = new StreamsSink(handler, 10, limiter)
+    const sinkP = sink.start(ctrl.signal)
+
+    const checkpoint = vi.fn().mockResolvedValue(undefined)
+
+    const batch = (from: number, count = 10): IngesterBatch => ({
+      items: Array.from({ length: count }, (_, i) => [
+        StreamName.parse("Cat-stream0"),
+        mkEvent("Something", BigInt(from + i)),
+      ]),
+      isTail: false,
+      checkpoint: BigInt(from + count),
+      onComplete: checkpoint,
+    })
+
+    // each mkBatch has 1 event in 10 streams
+    await sink.pump(batch(0), ctrl.signal)
+    await limiter.waitForEmpty()
+    ctrl.abort()
+
+    const expectedIndices = [0n, 1n, 2n, 3n, 4n, 5n, 6n, 7n, 8n, 9n]
+    expect(eventIndices).toEqual(expectedIndices.concat(expectedIndices))
+    expect(invocations).toBe(2)
+    expect(checkpoint).toHaveBeenCalledTimes(1)
+    await sinkP
+  })
 })
